extract fileGetter from files controller
