+++ controllers/files.js
@@ -16,6 +16,7 @@ const shortid = require('shortid');
 const upload = multer({storage: multer.memoryStorage()});
 const _ = require('lodash');
 const winston = require('winston');
+const {changeQueryParams, getStorageContext, checkIfOfficeFiles, FileGetter} = require('../helpers/files/fileGetter');
 
 const filterOptions = [
     {key: 'pics', label: 'Bilder'},
@@ -77,23 +78,7 @@ const requestSignedUrl = (req, data) => {
     });
 };
 
-/**
- * handles query params for file requests
- */
-const changeQueryParams = (originalUrl, params = {}, pathname = '') => {
-    const urlParts = url.parse(originalUrl, true);
 
-    Object.keys(params).forEach(param => {
-        urlParts.query[param] = params[param];
-    });
-
-    if (pathname) {
-        urlParts.pathname = pathname;
-    }
-
-    delete urlParts.search;
-    return url.format(urlParts);
-};
 
 /**
  * generates the displayed breadcrumbs on the actual file page
@@ -121,70 +106,9 @@ const getBreadcrumbs = (req, {dir = '', baseLabel = '', basePath = '/files/my/'}
     return breadcrumbs;
 };
 
-/**
- * generates the correct file's or directory's storage context for further requests
- */
-const getStorageContext = (req, res, options = {}) => {
 
-    if (req.query.storageContext) {
-        return pathUtils.normalize(req.query.storageContext + '/');
-    }
 
-    let currentDir = options.dir || req.query.dir || '/';
-    const urlParts = url.parse((options.url || req.originalUrl), true);
 
-    let storageContext = urlParts.pathname.replace('/files/', '/');
-
-    if (storageContext === '/my/') {
-        storageContext = 'users/' + res.locals.currentUser._id + '/';
-    }
-
-    if (currentDir.slice(-1) !== '/') currentDir = currentDir + '/';
-    return pathUtils.join(storageContext, currentDir);
-};
-
-/**
- * fetches all files and directories for a given storageContext
- */
-const FileGetter = (req, res, next) => {
-    let path = getStorageContext(req, res);
-    let pathComponents = path.split('/');
-    if (pathComponents[0] === '') pathComponents = pathComponents.slice(1); // remove leading slash, if present
-    const currentDir = pathComponents.slice(2).join('/') || '/';
-
-    path = pathComponents.join('/');
-
-    return api(req).get('/fileStorage', {
-        qs: {path}
-    }).then(data => {
-        let {files, directories} = data;
-
-        files = files.map(file => {
-            file.file = file.key;
-            return file;
-        });
-
-        directories = directories.map(dir => {
-            const targetUrl = pathUtils.join(currentDir, dir.name);
-            dir.url = changeQueryParams(req.originalUrl, {dir: targetUrl});
-            dir.originalPath = path;
-            dir.path = pathUtils.join(path, dir.name);
-            return dir;
-        });
-
-        checkIfOfficeFiles(files);
-
-        res.locals.files = {
-            files,
-            directories,
-            path
-        };
-
-        next();
-    }).catch(err => {
-        next(err);
-    });
-};
 
 /**
  * fetches all sub-scopes (courses, classes etc.) for a given user and super-scope
@@ -255,29 +179,7 @@ const registerSharedPermission = (userId, filePath, shareToken, req) => {
     });
 };
 
-/**
- * check whether given files can be opened in LibreOffice
- */
-const checkIfOfficeFiles = files => {
-    if (!process.env.LIBRE_OFFICE_CLIENT_URL) {
-        logger.error('LibreOffice env is currently not defined.');
-        return;
-    }
 
-    const officeFileTypes = [
-        'application/vnd.openxmlformats-officedocument.wordprocessingml.document',     //.docx
-        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',           //.xlsx
-        'application/vnd.openxmlformats-officedocument.presentationml.presentation',   //.pptx
-        'application/vnd.ms-powerpoint',                                               //.ppt
-        'application/vnd.ms-excel',                                                    //.xlx
-        'application/vnd.ms-word',                                                     //.doc
-        'application/vnd.oasis.opendocument.text',                                     //.odt
-        'text/plain',                                                                  //.txt
-        'application/msword'                                                           //.doc
-    ];
-
-    files.forEach(f => f.isOfficeFile = officeFileTypes.indexOf(f.type) >= 0);
-};
 
 /**
  * generates the correct LibreOffice url for (only) opening office-files
@@ -619,19 +521,19 @@ router.get('/courses/:courseId', FileGetter, function (req, res, next) {
         if (['Schüler', 'Demo'].includes(res.locals.currentRole))
             canCreateFile = false;
 
-        res.render('files/files', Object.assign({
-            title: 'Dateien',
-            canUploadFile: true,
-            canCreateDir: true,
-            canCreateFile,
-            path: res.locals.files.path,
-            inline: req.query.inline || req.query.CKEditor,
-            CKEditor: req.query.CKEditor,
-            breadcrumbs,
-            showSearch: true,
-            courseId: req.params.courseId,
-            courseUrl: `/courses/${req.params.courseId}/`
-        }, res.locals.files));
+            res.render('files/files', Object.assign({
+                title: 'Dateien',
+                canUploadFile: true,
+                canCreateDir: true,
+                canCreateFile,
+                path: res.locals.files.path,
+                inline: req.query.inline || req.query.CKEditor,
+                CKEditor: req.query.CKEditor,
+                breadcrumbs,
+                showSearch: true,
+                courseId: req.params.courseId,
+                courseUrl: `/courses/${req.params.courseId}/`
+            }, res.locals.files));
 
     });
 });
@@ -842,7 +744,7 @@ router.post('/studentCanEdit', function(req, res, next) {
    })
        .then(_ => {
            res.json({success: true});
-       })
+       });
 });
